refactor controlled input to separate component
